CSS ring modifications.
Removed settings from title.css that were previously abstracted to style.css(for multiple .css files to use).  Commented out counter text and all html for the title screen to begin work on the ring with the same viewing area.

Created the floor, scoreboard, background area and ring poles for the ring.